W3 - Lazy Learning/requirements.txt
numpy~=1.21.6
scipy~=1.7.3
pandas~=1.1.5
sklearn~=0.0.post1
scikit-learn~=1.0.2
tabulate~=0.9.0
Orange3~=3.31.1